backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
cryptography==44.0.0
python-multipart==0.0.20

# HTTP client (for fetching swagger specs)
httpx==0.28.1

# Settings
pydantic-settings==2.7.1
pydantic[email]==2.10.5

# Utilities
python-dotenv==1.0.1
